+++ service/src/io/pedestal/service/protocols.clj
@@ -18,7 +18,7 @@
   is not a dependency available to the pedestal.service module."
 
   (response-buffer-size [this]
-    "Returns the buffer size in bytes available to this response, or nil if it can not be determined."))
+    "Returns the buffer size, in bytes, available to this response, or nil if it can not be determined."))
 
 ;; Cover the case where the servlet response object is not available.
 
@@ -35,7 +35,7 @@
   The connector, once started, will handle incoming requests and outgoing responses, building
   on the interceptors and initial context provided in the service map."
 
-  (start-connector [this]
+  (start-connector! [this]
     "Starts (or restarts) the connector.  If the :join? key of the service map is true, this method will block
     until the connector is stopped.
 
@@ -43,8 +43,17 @@
 
     Returns the connector.")
 
-  (stop-connector [this]
+  (stop-connector! [this]
     "Stops the connector, if started. Does nothing if not started.  This will unblock a thread
     that is blocked in start-connector.
 
-    Returns the connector."))
+    Returns the connector.")
+
+  (test-request [this ring-request]
+    "Test a Ring request map, returning a Ring response map.
+
+    This method will operate regardless of whether the connector is started or stopped.
+
+    The :body of the request must be nil, a String, a File, or an InputStream; likewise,
+    in the returned response the :body will be nil or an InputStream."))
+